project/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
Pillow==10.2.0
opencv-python==4.9.0.80
python-jose==3.3.0